# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.AnswerQuestionsRequestDTO:
    properties:
      answer_id:
        type: integer
      question_id:
        type: integer
      question_number:
        type: integer
    type: object
...
  dtos.CreateOptionDTO:
    properties:
      display_order:
        type: integer
      id:
        description: Nullable for creation
        type: integer
      is_active:
        type: boolean
      option_text:
        type: string
    type: object
  dtos.CreateQuestionDTO:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      language_id:
        type: integer
      options:
        items:
          $ref: '#/definitions/dtos.CreateOptionDTO'
        type: array
      ques_point:
        type: integer
      question_text:
        type: string
    type: object
  dtos.CreateQuestionsRequestDTO:
    properties:
      questions:
        items:
          $ref: '#/definitions/dtos.CreateQuestionDTO'
        type: array
    type: object
...
  dtos.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  dtos.GetQuestionByTextRequestDTO:
    properties:
      language_id:
        type: integer
      question_text:
        type: string
    required:
    - language_id
    - question_text
    type: object
...
  dtos.OptionDTO:
    properties:
      display_order:
        type: integer
      id:
        type: integer
      option_text:
        type: string
    type: object
...
  dtos.QuestionResponse:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      language_id:
        type: integer
      question_text:
        type: string
    type: object
  dtos.QuestionResponseDTO:
    properties:
      id:
        type: integer
      language_id:
        type: integer
      options:
        items:
          $ref: '#/definitions/dtos.OptionDTO'
        type: array
      ques_point:
        type: integer
      question_text:
        type: string
      selected_option:
        type: integer
    type: object
  dtos.QuestionSubmitRequest:
    properties:
      language_id:
        type: integer
      question_text:
        type: string
    required:
    - language_id
    - question_text
    type: object
...
  dtos.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@thumsup.com
    name: API Support
  description: API documentation for Thums Up Backend Service including Contest Week
    Management
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Thums Up Backend API
  version: "1.0"
paths:
...
  /profile/questions:
    get:
      consumes:
      - application/json
      description: Get all active questions with user's answers
      parameters:
      - description: Language ID
        in: query
        name: language_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Questions retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.QuestionResponseDTO'
                  type: array
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to fetch questions
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - Bearer: []
      summary: Get Questions
      tags:
      - Questions
    post:
      consumes:
      - application/json
      description: Submit answers to multiple questions
      parameters:
      - description: Answer Questions Request
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/dtos.AnswerQuestionsRequestDTO'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Answers submitted successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to submit answers
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - Bearer: []
      summary: Answer Questions
      tags:
      - Questions
  /profile/questions/create:
    post:
      consumes:
      - application/json
      description: Create or update questions and options
      parameters:
      - description: Create Questions Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateQuestionsRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Questions created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to create questions
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - Bearer: []
      summary: Create Questions
      tags:
      - Questions
  /profile/questions/text:
    post:
      consumes:
      - application/json
      description: Get a question by its text and language ID
      parameters:
      - description: Get Question By Text Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.GetQuestionByTextRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Question retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.QuestionResponseDTO'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Question not found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - Bearer: []
      summary: Get Question By Text
      tags:
      - Questions
  /questions:
    post:
      consumes:
      - application/json
      description: Submit a new question with text and language. Requires authentication.
      parameters:
      - description: Question text and language ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.QuestionSubmitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Question submitted successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.QuestionResponse'
              type: object
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Failed to submit question
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - Bearer: []
      summary: Submit a new question
      tags:
      - Questions
  /questions/active:
    get:
      consumes:
      - application/json
      description: Retrieve all active questions
      produces:
      - application/json
      responses:
        "200":
          description: Questions retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.QuestionResponse'
                  type: array
              type: object
        "500":
          description: Failed to get active questions
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Get active questions
      tags:
      - Questions
...
securityDefinitions:
  APIKey:
    description: Admin API key for protected operations
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"